+++ build_ext.py
@@ -28,6 +28,7 @@ def build(setup_kwargs: Any) -> None:
                         "src/zeroconf/_listener.py",
                         "src/zeroconf/_protocol/incoming.py",
                         "src/zeroconf/_protocol/outgoing.py",
+                        "src/zeroconf/_services/registry.py",
                     ],
                     compiler_directives={"language_level": "3"},  # Python 3
                 ),